requirements.txt
streamlit
requests
langchain-cohere
mcp
langchain-community
phidata
pycountry
googlesearch-python
uv
langchain-mcp-adapters